requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
astroid==3.0.0
asttokens==2.4.0
autopep8==2.0.4
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.3
executing==2.0.0
Faker==19.8.0
fastapi==0.103.1
fire==0.5.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
icecream==2.1.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
numpy==1.26.0
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
pylint==3.0.1
pymongo==4.5.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
termcolor==2.3.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
